requirements.txt
langchain
llama-cpp-python
PyPDF2==3.0.1
faiss-cpu==1.7.4
ctransformers
pypdf
chromadb
tiktoken
streamlit-extras
InstructorEmbedding
sentence-transformers
jq
langchain-community
openai